requirements.txt
dateparser==0.7.2
jsonlines==1.2.0
pandas==0.25.3
python-slugify==4.0.0
Scrapy==1.8.0

